+++ sdmx-dl-provider-px/src/main/java/module-info.java
@@ -1,0 +1,19 @@
+import sdmxdl.provider.px.drivers.PxWebDriver;
+import sdmxdl.web.spi.Driver;
+
+module sdmxdl.provider.px {
+
+    requires static lombok;
+    requires static nbbrd.design;
+    requires static nbbrd.service;
+    requires static org.checkerframework.checker.qual;
+
+    requires sdmxdl.api;
+    requires sdmxdl.format.xml;
+    requires sdmxdl.provider.ri;
+    requires sdmxdl.provider.base;
+    requires com.google.gson;
+
+    provides Driver with
+            PxWebDriver;
+}