+++ CHANGELOG.md
@@ -28,6 +28,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - ![SOURCE] Update ECB endpoint [#495](https://github.com/nbbrd/sdmx-dl/issues/495)
 - ![FORMAT] Refactor file format API [#502](https://github.com/nbbrd/sdmx-dl/issues/502)
 - ![PROVIDER] Move curl backend to an external project
+- ![PROVIDER] Move PxWebDriver to its own module
 - ![CLI] Use environment variables as default values
 - ![GRPC] Migrate gRPC module to Quarkus framework